requirements-dev.txt
# everything the developer needs in addition to the production requirements

# testing
pytest-mock
pytest>=5.2
pytest-flask
mongomock==3.20
responses

# utils
invoke
black

# docs
mkdocs
mkdocs-material
